@@ -7,42 +7,39 @@
     "build": "webpack"
   },
   "dependencies": {
-    "@babel/standalone": "^7.13.13",
-    "@codemirror/autocomplete": "^0.19.0",
-    "@codemirror/closebrackets": "^0.19.0",
-    "@codemirror/commands": "^0.19.0",
-    "@codemirror/comment": "^0.19.0",
-    "@codemirror/gutter": "^0.19.0",
-    "@codemirror/highlight": "^0.19.0",
-    "@codemirror/history": "^0.19.0",
-    "@codemirror/lang-javascript": "^0.19.0",
-    "@codemirror/language": "^0.19.0",
-    "@codemirror/search": "^0.19.0",
-    "@codemirror/state": "^0.19.0",
-    "@codemirror/view": "^0.19.0",
-    "@expressive/babel-preset-react": "^1.0.0-beta.94",
-    "@expressive/css": "^1.0.0-beta.90",
-    "@expressive/layout": "^1.0.0-beta.7",
-    "@expressive/mvc": "^1.0.0-beta.27",
-    "prettier": "^2.1.2",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0"
+    "@babel/standalone": "^7.22.4",
+    "@codemirror/autocomplete": "^6.7.1",
+    "@codemirror/commands": "^6.2.4",
+    "@codemirror/lang-javascript": "^6.1.9",
+    "@codemirror/language": "^6.7.0",
+    "@codemirror/search": "^6.4.0",
+    "@codemirror/state": "^6.2.1",
+    "@codemirror/view": "^6.12.0",
+    "@expressive/babel-preset-react": "^0.1.0",
+    "@expressive/css": "^0.1.0",
+    "@expressive/react": "^0.28.0",
+    "@lezer/common": "^1.0.3",
+    "@lezer/highlight": "^1.1.6",
+    "@types/react-dom": "^18.2.4",
+    "prettier": "^2.8.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
-    "@types/react": "^17.0.19",
-    "babel-loader": "^8.2.2",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "react-refresh": "^0.10.0",
-    "style-loader": "^3.2.1",
-    "webpack": "^5.52.1",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.2.0"
+    "@babel/core": "^7.22.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-react": "^7.22.3",
+    "@babel/preset-typescript": "^7.21.5",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
+    "@types/react": "^18.2.8",
+    "babel-loader": "^9.1.2",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
+    "html-webpack-plugin": "^5.5.1",
+    "react-refresh": "^0.14.0",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.85.0",
+    "webpack-cli": "^5.1.1",
+    "webpack-dev-server": "^4.15.0"
   }
 }
