@@ -1,21 +1,33 @@
 import { Component } from "react";
 
 class Boundary extends Component {
-  state = {};
+  state = {
+    hasError: false,
+    error: { message: '', stack: '' },
+    info: { componentStack: '' }
+  };
 
-  static getDerivedStateFromError(error){
-    console.error(error);
-    return {
-      error: "Something went wrong while rendering."
-    };
-  }
+  static getDerivedStateFromError = error => {
+    return { hasError: true };
+  };
 
-  componentDidCatch(err){
-    this.props.onError(err);
-  }
+  // componentDidCatch = (error, info) => {
+  //   debugger;
+  //   this.setState({ error, info, hasError: true });
+  // };
 
-  render(){
-    return this.state.error || this.props.children;
+  render() {
+    const { hasError, error, info } = this.state;
+    const { children } = this.props;
+
+    if(hasError)
+      <div>There was an error</div>
+    else
+      <this>{children}</this>
+
+    // return hasError
+    //   ? <div>There was error or something.</div>
+    //   : children;
   }
 }
 
